--- a/branca_test.go
+++ b/branca_test.go
@@ -1,172 +1,100 @@
 package branca
 
 import (
+	"encoding/hex"
 	"testing"
+	"time"
 )
 
-var (
-	testVectors []struct {
-		key       string
-		nonce     string
-		timestamp uint32
-		payload   string
-		expected  string
-	}
-)
+var testKey = func() (key [32]byte) {
+	copy(key[:], []byte("supersecretkeyyoushouldnotcommit"))
+	return
+}()
 
 // TestVector1 for testing encoding data to a valid branca token.
 func TestVector1(t *testing.T) {
-	testVectors = []struct {
-		key       string
+	tests := []struct {
+		key       [32]byte
 		nonce     string
-		timestamp uint32
+		timestamp time.Time
 		payload   string
 		expected  string
-	}{
-		{"supersecretkeyyoushouldnotcommit", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a"},
-	}
+	}{{
+		key:       testKey,
+		nonce:     "0102030405060708090a0b0c0102030405060708090a0b0c",
+		timestamp: time.Unix(0, 0).Add(time.Second * 123206400),
+		payload:   "Hello world!",
+		expected:  "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a",
+	}}
 
-	for _, table := range testVectors {
-		b := NewBranca(table.key)
-		b.setNonce(table.nonce)
-		b.setTimeStamp(table.timestamp)
+	for _, test := range tests {
+		b := New(test.key)
+		nonce := decodeHexNonce(test.nonce)
+		// Encode token.
+		encoded := b.EncodeToRawAtTime([]byte(test.payload), nonce, test.timestamp)
+		encodedStr := Base62(encoded)
+		if encodedStr != test.expected {
+			t.Errorf("EncodeToString(%q) = %s. got %s, expected %q", test.payload, encoded, encoded, test.expected)
+		}
 
-		// Encode string.
-		encoded, err := b.EncodeToString(table.payload)
+		// Decode token..
+		decoded, timestamp, err := b.Decode(encodedStr)
 		if err != nil {
 			t.Errorf("%q", err)
 		}
-		if encoded != table.expected {
-			t.Errorf("EncodeToString(\"%s\") = %s. got %s, expected %q", table.payload, encoded, encoded, table.expected)
+		if string(decoded) != test.payload {
+			t.Errorf("DecodeToString(%q) = %s. got %s, expected %q", test.expected, decoded, decoded, test.expected)
 		}
-
-		// Decode string.
-		decoded, err := b.DecodeToString(encoded)
-		if err != nil {
-			t.Errorf("%q", err)
-		}
-		if decoded != table.payload {
-			t.Errorf("DecodeToString(\"%s\") = %s. got %s, expected %q", table.expected, decoded, decoded, table.expected)
-		}
-	}
-}
-
-// TestVector2 for testing encoding data to a valid branca token with a TTL.
-func TestVector2(t *testing.T) {
-	testVectors = []struct {
-		key       string
-		nonce     string
-		timestamp uint32
-		payload   string
-		expected  string
-	}{
-		{"supersecretkeyyoushouldnotcommit", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a"},
-	}
-
-	for _, table := range testVectors {
-		b := NewBranca(table.key)
-		b.setNonce(table.nonce)
-		b.setTimeStamp(table.timestamp)
-
-		// Encode string.
-		encoded, err := b.EncodeToString(table.payload)
-		if err != nil {
-			t.Errorf("%q", err)
-		}
-		if encoded != table.expected {
-			t.Errorf("EncodeToString(\"%s\") = %s. got %s, expected %q", table.payload, encoded, encoded, table.expected)
-		}
-
-		// Decode string with TTL. Should throw an error with no token encoded because it has expired.
-		b.SetTTL(3600)
-		decoded, derr := b.DecodeToString(encoded)
-		if derr == nil {
-			t.Errorf("%q", derr)
-		}
-		if decoded != "" {
-			t.Errorf("DecodeToString(\"%s\") = %s. got %s, expected %q", table.expected, decoded, decoded, table.expected)
-		}
-	}
-}
-
-// TestGenerateToken for testing issuing branca tokens.
-func TestGenerateToken(t *testing.T) {
-	testVectors = []struct {
-		key       string
-		nonce     string
-		timestamp uint32
-		payload   string
-		expected  string
-	}{
-		{"supersecretkeyyoushouldnotcommit", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a"},
-	}
-
-	for _, table := range testVectors {
-		// Not generated with set timestamp.
-		b := NewBranca(table.key)
-
-		// Encode string.
-		encoded, err := b.EncodeToString(table.payload)
-		if err != nil {
-			t.Errorf("%q", err)
-		}
-		if encoded == table.expected {
-			t.Errorf("EncodeToString(\"%s\") = %s. got %s, expected %q", table.payload, encoded, encoded, table.expected)
-		}
-	}
-}
-
-// TestInvalidEncodeString for testing errors when generating branca tokens.
-func TestInvalidEncodeString(t *testing.T) {
-	testVectors = []struct {
-		key       string
-		nonce     string
-		timestamp uint32
-		payload   string
-		expected  string
-	}{
-		{"supersecretkeyyoushouldnotcommi", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a"}, // Invalid key
-
-		{"supersecretkeyyoushouldnotcommi", "", 123206400, "Hello world!",
-			"875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a"}, // Invalid key + no nonce
-
-	}
-
-	for _, table := range testVectors {
-		b := NewBranca(table.key)
-
-		_, err := b.EncodeToString(table.payload)
-		if err == nil {
-			t.Errorf("%q", err)
+		if !timestamp.Equal(test.timestamp) {
+			t.Errorf("unexpected decoded timestamp; got %v want %v", timestamp, test.timestamp)
 		}
 	}
 }
 
 // TestInvalidDecodeString for testing errors when decoding branca tokens.
 func TestInvalidDecodeString(t *testing.T) {
-	testVectors = []struct {
-		key       string
-		nonce     string
-		timestamp uint32
-		payload   string
-		expected  string
-	}{
-		{"supersecretkeyyoushouldnotcommit", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0"}, // Invalid base62
-
-		{"supersecretkeyyoushouldnotcommi", "", 123206400, "Hello world!",
-			"875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsA"}, // Invalid key + Invalid base62.
-
-		{"supersecretkeyyoushouldnotcommi", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a"}, // Invalid key
-
-		{"supersecretkeyyoushouldnotcommit", "0102030405060708090a0b0c0102030405060708090a0b0c", 123206400, "Hello world!", "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLOZtQ0ekPHt8kJHQp0a"}, // Invalid malformed base62
+	tests := []struct {
+		testName    string
+		key         [32]byte
+		token       string
+		expectError string
+	}{{
+		testName:    "invalid-base62",
+		key:         testKey,
+		token:       "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a-",
+		expectError: "invalid base62 token",
+	}, {
+		testName:    "invalid-key",
+		key:         [32]byte{},
+		token:       "875GH233T7IYrxtgXxlQBYiFobZMQdHAT51vChKsAIYCFxZtL1evV54vYqLyZtQ0ekPHt8kJHQp0a",
+		expectError: "chacha20poly1305: message authentication failed",
+	}, {
+		testName:    "data-too-short",
+		token:       "875GH233T7IYrxtgXxlQBYiFo",
+		expectError: "invalid base62 token",
+	}}
+	for _, test := range tests {
+		t.Run(test.testName, func(t *testing.T) {
+			b := New(test.key)
+			_, _, err := b.Decode(test.token)
+			if err == nil {
+				t.Fatalf("unexpected success")
+			}
+			if err.Error() != test.expectError {
+				t.Fatalf("unexpected error; got %q want %q", err, test.expectError)
+			}
+		})
 	}
+}
 
-	for _, table := range testVectors {
-		b := NewBranca(table.key)
-
-		_, err := b.DecodeToString(table.expected)
-		if err == nil {
-			t.Errorf("%q", err)
-		}
+func decodeHexNonce(s string) (nonce [24]byte) {
+	data, err := hex.DecodeString(s)
+	if err != nil {
+		panic(err)
 	}
+	if len(data) != len(nonce) {
+		panic("bad length")
+	}
+	copy(nonce[:], data)
+	return
 }