--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,8 @@
+module github.com/CanonicalLtd/branca
+
+go 1.11
+
+require (
+	golang.org/x/crypto v0.0.0-20181203042331-505ab145d0a9
+	golang.org/x/sys v0.0.0-20181205085412-a5c9d58dba9a // indirect
+)