--- a/internal/basex/basex.go
+++ b/internal/basex/basex.go
@@ -1,22 +1,30 @@
 // Package basex provides fast base encoding / decoding of any given alphabet using bitcoin style leading zero compression.
 // It is a GO port of https://github.com/cryptocoinjs/base-x
+//
+// It has been copied from github.com/eknkc/basex
+// to avoid that external dependency, and updated somewhat
+// for efficiency.
 package basex
 
 import (
-	"bytes"
 	"errors"
+	"strings"
 )
 
 // Encoding is a custom base encoding defined by an alphabet.
 // It should bre created using NewEncoding function
 type Encoding struct {
-	base        int
-	alphabet    []rune
-	alphabetMap map[rune]int
+	base     int
+	alphabet string
+	// alphabetMap holds a mapping from alphabet character (index into alphabetMap)
+	// to 1 + digit encoded by that alphabet character.
+	alphabetMap [128]byte
+	zero        byte
 }
 
 // NewEncoding returns a custom base encoder defined by the alphabet string.
 // The alphabet should contain non-repeating characters.
+// It does not allow non-ASCII characters in the alphabet.
 // Ordering is important.
 // Example alphabets:
 //   - base2: 01
@@ -24,22 +32,28 @@ type Encoding struct {
 //   - base32: 0123456789ABCDEFGHJKMNPQRSTVWXYZ
 //   - base62: 0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ
 func NewEncoding(alphabet string) (*Encoding, error) {
-	runes := []rune(alphabet)
-	runeMap := make(map[rune]int)
-
-	for i := 0; i < len(runes); i++ {
-		if _, ok := runeMap[runes[i]]; ok {
-			return nil, errors.New("Ambiguous alphabet.")
-		}
-
-		runeMap[runes[i]] = i
+	if len(alphabet) == 0 {
+		return nil, errors.New("empty alphabet")
 	}
-
-	return &Encoding{
-		base:        len(runes),
-		alphabet:    runes,
-		alphabetMap: runeMap,
-	}, nil
+	if len(alphabet) > 128 {
+		return nil, errors.New("alphabet too large")
+	}
+	e := &Encoding{
+		base:     len(alphabet),
+		alphabet: alphabet,
+		zero:     alphabet[0],
+	}
+	for i := 0; i < len(alphabet); i++ {
+		b := alphabet[i]
+		if b >= 128 {
+			return nil, errors.New("out of range alphabet character")
+		}
+		if e.alphabetMap[b] > 0 {
+			return nil, errors.New("duplicate character found in alphabet")
+		}
+		e.alphabetMap[b] = byte(i) + 1
+	}
+	return e, nil
 }
 
 // Encode function receives a byte slice and encodes it to a string using the alphabet provided
@@ -49,30 +63,25 @@ func (e *Encoding) Encode(source []byte) string {
 	}
 
 	digits := []int{0}
-
 	for i := 0; i < len(source); i++ {
 		carry := int(source[i])
-
 		for j := 0; j < len(digits); j++ {
 			carry += digits[j] << 8
 			digits[j] = carry % e.base
 			carry = carry / e.base
 		}
-
 		for carry > 0 {
 			digits = append(digits, carry%e.base)
 			carry = carry / e.base
 		}
 	}
 
-	var res bytes.Buffer
-
+	var res strings.Builder
 	for k := 0; source[k] == 0 && k < len(source)-1; k++ {
-		res.WriteRune(e.alphabet[0])
+		res.WriteByte(e.zero)
 	}
-
 	for q := len(digits) - 1; q >= 0; q-- {
-		res.WriteRune(e.alphabet[digits[q]])
+		res.WriteByte(e.alphabet[digits[q]])
 	}
 
 	return res.String()
@@ -85,31 +94,31 @@ func (e *Encoding) Decode(source string) ([]byte, error) {
 		return []byte{}, nil
 	}
 
-	runes := []rune(source)
-
 	bytes := []byte{0}
 	for i := 0; i < len(source); i++ {
-		value, ok := e.alphabetMap[runes[i]]
-
-		if !ok {
-			return nil, errors.New("Non Base Character")
+		c := source[i]
+		if int(c) >= len(e.alphabetMap) {
+			return nil, errors.New("unexpected character found")
 		}
-
-		carry := int(value)
+		value := int(e.alphabetMap[c])
+		if value == 0 {
+			return nil, errors.New("unexpected character found")
+		}
+		value--
 
 		for j := 0; j < len(bytes); j++ {
-			carry += int(bytes[j]) * e.base
-			bytes[j] = byte(carry & 0xff)
-			carry >>= 8
+			value += int(bytes[j]) * e.base
+			bytes[j] = byte(value & 0xff)
+			value >>= 8
 		}
 
-		for carry > 0 {
-			bytes = append(bytes, byte(carry&0xff))
-			carry >>= 8
+		for value > 0 {
+			bytes = append(bytes, byte(value&0xff))
+			value >>= 8
 		}
 	}
 
-	for k := 0; runes[k] == e.alphabet[0] && k < len(runes)-1; k++ {
+	for k := 0; source[k] == e.zero && k < len(source)-1; k++ {
 		bytes = append(bytes, 0)
 	}