--- a/internal/basex/go.mod
+++ b/internal/basex/go.mod
@@ -1,1 +1,0 @@
-module github.com/eknkc/basex