--- a/internal/fastuuid/go.mod
+++ b/internal/fastuuid/go.mod
@@ -1,3 +1,0 @@
-module github.com/rogpeppe/fastuuid
-
-go 1.12