… .gitignore
… .travis.yml
… go.sum
… internal/basex/LICENSE
… internal/basex/README.md
… internal/fastuuid/LICENSE
… internal/fastuuid/README.md