This updates the Branca API and code after forking from `github.com/hako/branca`.

Specifically:

 - the token payload is now `[]byte` not `string`.
 - the client is responsible for checking timestamp expiry, which simplifies the API and makes it possible to use a different expiry duration based on the payload contents.
 - it's possible for external code to choose the nonce and timestamp, which is useful for tests.
 - the external basex dependency has been removed.
 - encoding a token no longer returns an error.
 - the base62 wrapping is now optional, making it efficient to fold a token into some other binary format without incurring double-encoding cost.
- the tests check the actual error returns (the old tests were not testing what they intended to test)

We vendor two small external packages to avoid external
unnecessary external dependencies - these were copied in
before this PR so this PR shows all the changes that
have been made.

Benchmark of old vs new code:

```
benchmark             old ns/op     new ns/op     delta
BenchmarkEncode-4     44003         35954         -18.29%
BenchmarkDecode-4     12130         1975          -83.72%
```

Note that in the encoding, ~95% of the time is spent in base62.Encode,
so there is room for improvement (and only a limited amount
of room for improvement without changing that code).